requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-nvidia-ai-endpoints>=0.3
langgraph>=0.2
tavily-python>=0.5
python-dotenv>=1.0
